Add TLS support for ZooKeeper
This adds a script to generate TLS certs for zookeeper.

It also adds new config file options for specifying certs for a
TLS connection, adds a howto document to advise admins on how
to configure ZK for TLS.

It also removes the 'required' flag for the SASL auth parameters,
since they are not actually required.

Change-Id: Icd976cc32dfd9f75f8cfb1c9ad11e08af31723d6