Don't access parent layout errors if there is no parent layout
There are situations (when the parent layout is in error?) where the
parent_layout is None when looking to determine if our current layout
errors are related to the current change or not.

Handle this by checking the delta between parent_layout and current
layout only when parent_layout is not None. Then assume any errors in
the layout are related to the current change if there is no parent
layout so that they will get reported.

Change-Id: Ic5985870f5642496abbc5d0ef400be4ea6b4a74a