Use buildkit to mount instead of copy /output
We currently copy /output into the target image so we can do the
install of the artifacts. Although we then delete it, it's still
in that intermediate layer. That is, of course, silly. Instead,
use the buildkit extensions to mount /output from the builder
image.

We could probably remove the copy if we could update the python-builder
install-from-bindep command to understand relative paths.

Change-Id: Ie0b2b92bbf712a259b3e69291268bf08c37f89df