Defer setting build pause to event queue
We currently set the paused flag on the build object directly when we
receive the notification from the executor and afterwards enqueue a
BuildPaused event to let the pipeline process the pause. As soon a the
build is marked as paused the pipeline processor will request
resources for the child jobs. However during processing of the
BuildPaused event it can be that the paused job requested to skip some
jobs. In this case we have a race between requesting the resources and
setting the child job results to SKIPPED. The result is that if the
resources are requested earlier they will be leaked because they will
never be released by the scheduler as no builds will be enqueued for
them.

This can be solved by deferring to set the paused flag to the event
queue which eliminates the race.

This is similar to the deferred result setting in
I0631b97e6ae2bf7d883c6c6a189e9d41d99ece5d.

Change-Id: Ie0fa8a2613fca65884d54b59273769f4984552cd