Don't accept nodes for unknown requests
In times when there is an event backlog in the scheduler it can happen
that _updateNodeRequest submits the NodesProvisioned event while there
is already an event in the queue that will result in deleting the node
request because it's obsolete. During acceptNodes we check for
request.canceled which is not necessarily changed alongside which
leads to a node leak in this case until the scheduler is
restarted. This can be fixed by also checking if we still know the
request we're accepting the nodes for.

Change-Id: I21604ad0bb3e5f4770302e3cf58899b35cf4e558