WIP: Enforce sql connections for scheduler and web
On the way towards a fully scale out scheduler we need to move the
times database from the local filesystem into the sql
database. Therefore we need to make at least one sql connection
mandatory.

In a later change we will mark one connection as primary which is used
for reporting regardless of the pipeline configuration.

Change-Id: I30723f9b320b9f2937cc1d7ff3267519161bc380
Depends-On: https://review.opendev.org/621479
Story: 2007192
Task: 38329