Update to create-react-app 3.4.1
While trying to figure out why the ansi patch wouldn't
build it emerged that we're now 2 major releases behind
on create-react-app.

Update create-react-app to the latest 3.4.1. This also updates
react to 16.13, and updates eslint globals processing so that
we don't have to declare globals in headers when we've
declared them already in the eslint file.

Two links were turned in to buttons. There is now a warning/error
about accessibility that considers an href-less a with an onclick
to be an anti-pattern for folks with accessiblity issues. Changing
it to a button is the recommended solution.

Finally, although this doesn't do it, create-react-app 3
has support for typescript, so if we want we can start migrating
files to .ts or .tsx extensions and start doing typing in
them.

Depends-On: https://review.opendev.org/726546
Change-Id: Ibc69bef605a62e4fdd2ebba33d9d1b822e7dfeba