WIP: Store unparsed branch config in Zookeeper
We need to store the global configuration in zookeeper. However
zookeeper is not designed as a database with a large amount of data we
should store as little as possible in zookeeper. Thus we only store
the per project-branch unparsed config in zookeeper. From this every
part of zuul like the scheduler or also zuul-web can quickly
recalculate the layout of each tenant and keep it up to date by
watching for changes in the unparsed project-branch-config. We will
lock the complete global config with one lock and maintain a checkpoint
version of it. This way each component can watch the config version
number and react accordingly. Although we lock the complete global
config we still should store the actual config in distinct nodes per
project and branch. This is needed because of the 1MB limit per znode
in zookeeper. It further makes it less expensive to cache the global
config in each component as this cache will be updated incrementally.

The configs will be stored in the path
``/zuul/config/<project>/<branch>`` per branch segmented in
``/zuul/config/<project>/<branch>/<path-to-config>/<shard>``. The
``shard`` is a sequence number and will be used to store larger than
1MB files due to the limitation mentioned above.

Manual reconfiguration on one instance will reload the config files and
update zookeeper in case of differences. Other instances will be
notified about those changes via zookeeper watchers. During start
zookeeper will be used as primary store.

Change-Id: I519c4e6af1f21efdeed2b919575bcad9c50a9f18
Story: 2007192
Task: 38620