Update patch set 2
Patch Set 2: Code-Review-1

Having all branches share a queue was a very deliberate design decision for Zuul.  It's critical for any project that wants to be able to upgrade.  Most of the time, when people ask for this, and we talk to them, they haven't realized that merging changes to different branches out of order could break upgrade testing.

Is your use case one where you never have to worry about upgrades from previous releases?  If that's the case, then I think separating queues by branch could be okay.  But considering the importance of gating (and how easy it is to get wrong), this should not be the default, and it's only something that should be used by someone who fully understands the implications.  To take the OpenStack use case: we know it's the wrong choice for OpenStack, but not all of our developers do, so we need to prevent people from being able to do this.  Also, with the current implementation, it seems that if most of the project pipelines were defined in a config project, it could be very easy to accidentally end up with more queues than desired just by adding an in-repo project pipeline config.  If we're going to proceed with this, we'll have to address these points.

If, on the other hand, you do need to consider upgrades between branches (or there is any other kind of a relationship between them), and the issue is just that changes that you deem more important are sitting behind changes deemed less important in a queue, then maybe we could brainstorm some other ideas to address that?  Perhaps you could make more use of the promote feature?  Or perhaps we could look into adding some support for re-ordering queues on gate resets (so if an important item is behind an item that restarts, it gets automatically promoted)?

Patch-set: 2
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1