Update patch set 2
Patch Set 2:

I think we could do this fairly easily in the tenant, config, but I don't think it *needs* to be in the tenant config; it might be worth seeing if we can do this entirely in-repo.  Maybe put that setting in the project stanza, default it to shared-branches, and if a site wants to force the setting, they can do it in a config-project.

Another issue worth thinking about is whether this should be implicit or explicit.  For example, if two projects share a change queue, does each automatically have a shared queue per branch?  or could they share some project-branch queues but not others?  This could be accomplished by naming the queues differently in each branch (eg: "queue: shared-multibranch" and "queue: shared-stable-diablo").  Finally, what happens if two projects have differing settings (eg, nova has upgrade testing but swift does not)?  That might argue for using a tenant-wide setting after all.

Patch-set: 2