Move queue from pipeline to project
Referring to a queue makes more sense in the project where it has more
use such as allowing circular dependencies later. This makes it
possible to refer to the queue on project level which takes precedence
over referring queues in the pipeline config. We'll also deprecate
referring to the queue in the pipeline config and remove this in a
later release.

Change-Id: I14533ae57cfe688b55c26a1e17ab38e133180b28