Update patch set 10
Patch Set 10:

While I do support the idea of separated stream the reality is that it is very hard to get a decent implementation.

If you split them, you endup with line order mismatch between stdout/stderr, to the extend that makes   you wonder where is the error, or where does the error comes from. Much worse than a joined-stream, which has decent ordering


I would love to see this working but as someone that had to cope with the same kind of issue in several places, I doubt is achievable.

Ideally if we do this right, we should be able to display STDERR output colored in red, which would be great.

Patch-set: 10
Reviewer: Gerrit User 24162 <24162@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0