Add flag for disabling log streaming on a host
Some hosts don't run the log streaming daemon even though they
would otherwise be capable of doing so. Examples include hosts
added as part of a CD deployment. When jobs run on them, the
log streaming is filled with "waiting for log stream" messages
on every shell and command task.

Add a flag, zuul_skip_log_streaming that can be set to disable
zuul attempting to live stream content. This can be added when
such a host is added to the inventory.

Change-Id: Ie33a7caedb6c998bb11f8b013894ba517077eb1a