[wip] intermediate jobs
I recently had an issue using our nodepool container jobs from the dib
repository.  Nodepool jobs define a base container job and then two
variants; one for installing with released tools and one for
installing with siblings from git.  In this case, you never really
want to inherit from the base job; you want to choose one of the two
variants.  Although the base job was marked abstract, it didn't really
stop me accidentally inheriting from it directly.

This proposes an "intermediate" flag that would be applied to such a
base job.  In a case such as above, it would have raised an error and
alerted me to use one of the two other jobs.

Change-Id: Ifbb9ffa65f86a6b86b63a38e3234a12b564ba3c1