Make fake test Gerrit merger more realistic
The current Gerrit simulated merge operation simply resets the
target branch to the specified change's commit -- it doesn't actually
perform a git merge operation as gerrit (typically) would.  This
makes some test situations difficult, such as when we want to rely on
a post-merge pipeline preparing git repos on top of previously merged
changes.

This change actually performs git merge operations when gerrit changes
are merged in tests, so the result should be far more realistic.

Change-Id: Ia5b4b83096de3b8719c8b6d256e8e33749e7a5b8