Make console lines wrapping user configurable
Add a new option to the settings screen that allows user to control
how the long console (stdout/stderr) lines are displayed.

The new default beaviour is to perform soft-wrapping to the current
window, like normal terminals do.

Those that want to stick to the old behavior of keeping them on a
single line and use horizontal scrolling to read the line can uncheck
the option.

Screenshot: https://sbarnea.com/ss/Screen-Shot-2020-11-27-14-06-48.png
Change-Id: Ifa3a67346e86c1353a4446293a8a38acc02e0f0b