WIP: Revert "Revert "Tune automatic garbage collection of git repos""
TODO: check if this can be reproduced in a local test case

The reason for occasionally having invalid FETCH_HEADS has been
tracked to git's behavior of ignoring FETCH_HEAD when checking for
unrechable objects [1]. In order to solve this issue change the fetch
command to map the fetched ref to refs/zuul/fetch. This ref is unused
but will ensure that git won't prune the referenced object.

This reverts commit 9b300bc8dfc59584e3046337e7c001fc927d670f.

[1] https://public-inbox.org/git/20160708025948.GA3226@x/

Change-Id: Iea8d376b840a6486a9fa51b6654641d3b14f90f3