Do not add dependent changes project to zuul.projects
This causes problems when a project wants to start testing together with
another project in the tenant and a "Depends-On:" header is added to the
commit but forgetting to add a required-projects: property to the job.

The job succeeds since currently Depends-On changes are added to
zuul.projects but the following changes would fail since zuul would no
longer add that project to zuul.projects.

Change-Id: Iace8814a09fc84f948a60d1fce32398ee1af9acf