Update patch set 2
Patch Set 2: Code-Review-1

I'm not sure this is desirable.  One of the benefits of looking at logs in the web is that long lines are more easily horizontally scrolled.  This is useful, especially on test or debug log output, much of which has very long lines and is easier to read if they are not wrapped (because it is easier to scan down columnar output to find debug levels, timestamps, and the start of log messages.  Just because wrapping happens in some terminals doesn't mean it's the best solution.  For a counter-example, journalctl by default overcomes this limitation and implements horizontal scrolling so that it is easier to read log output.

Patch-set: 2
Label: Code-Review=-1