Don't exit zuul-web when signals are received
One of the last things the primary thread in zuul-web does after
configuring and starting the service is to call signal.pause(). This
causes it to wait for any signal to arrive and once a signal has arrived
and been handled we stopped zuul-web and killed the process.

This is a problem because SIGUSR2 is expected to toggle preformance and
debug tooling in the process. But since we exit after signal.pause()
returns we can't do that properly.

Fix this by running signal.pause() in a loop, but check if the web
object has stopped running and break out of the loop if so.

Change-Id: I66fbb86c25479823faeded8ede9c6204846c35c1