Fix loading_errors bug
This corrects an error
  Error in dynamic layout
  Traceback (most recent call last):
    File "/usr/local/lib/python3.8/site-packages/zuul/manager/__init__.py", line 625, in _loadDynamicLayout
      relevant_errors = self._findRelevantErrors(item,
    File "/usr/local/lib/python3.8/site-packages/zuul/manager/__init__.py", line 517, in _findRelevantErrors
      parent_layout.loading_errors.error_keys) or
  AttributeError: 'NoneType' object has no attribute 'loading_errors'

which can be caused when the following conditions are met:

* Tenant A is listed first.
* Tenant A holds projects A and B.
* Project B references a config object defined in project A.
* Tenant A loads without errors.
* Tenant B holds projects B and C.
* Tenant B has a standing configuration error due to the unknown
  reference to an object in project A from project B.
* Two or more changes to project C which cause dynamic
  configurations to be created are enqueued in a pipeline.
* The merge job for the second change finishes before the first.

This is because:

We cache configuration objects if they load without error in
their first tenant; that means that they can show up as errors in
later tenants, but as long as those other tenants aren't
proposing changes to that repo, it doesn't matter.  But it does
mean that every dynamic reconfiguration in this tenant will see
errors and will execute the code path that compares the new
dynamic configuration to the previous one to see if they are
relevant.

If a merge job for a dynamic config change arrives out of order,
we will compare it to the previous configuration to determine if
they are relevant, but since the previous layout had not been
calculated yet, the exception above was hit.

The solution is to indicate that the layout for the later change
is not ready until the layout for the previous change is.

Change-Id: Ibe1392a494d42b65080ab7e42f116db3869548ff