Add some debug lines to help with the loading_errors bug
I suspect that we're seeing this bug when a merge job for change
B arrives before change A, and change B creates a dynamic config.
However, a config error is still needed in order to trigger the
code path, and I've been unable to identify a likely candidate.
It would be useful to see what config errors are present in the
dynamic layout to learn how the bug is being triggered.

Change-Id: I2985fd8d69a971c959987896dae1da60533700eb