executor: Catch error when reading cpu_times
On older Linux Kernels (especially 3.10 shipped with CentOS 7), having
the mount option hidepid=2 set for /proc causes all zombie-processes to
vanish from /proc for non-root users. We rely on /proc to get ansible's
cpu_times however. Currently a fatal psutil.NoSuchProcess exception is
thrown in the given case.

Now, the error is catched and a warning is printed instead. This comes
with the drawback that cpu_time measuring does not work anymore (it will
always print 0s).

Another possible solution for this would be to read the cpu_times of
/proc/self before and after the execution of the ansible subprocess.
This is the behaviour that is present in GNU time for example. For the
described edge-case with an old kernel, this is considered overkill.

Change-Id: I776f8588e1b0f6f3d145b0c6c54c8970fe864251