Update patch set 20
Patch Set 20: Code-Review-1

(1 comment)

Thanks for catching this!

The purpose of the quick-start test job is to test the instructions in the quick-start documentation.  Every task in there should correspond to something that we instruct the user to do.  While the additions here are a great test of gating functionality, they look more like a unit test than something we would ask a user to do.

I think that we should either update this patch to only include the change to the zuul.conf file (without updates to the playbooks), or (and I like this better) implement your default value suggestion (or maybe make it a required value).

Patch-set: 20
Label: Code-Review=-1