Update patch set 5
Patch Set 5:

Please don't hate me.  I think 'dequeue' might be better.  It is slightly confusing because successful items are dequeued, but I think that will be easier to understand.  Now that I revisited this with fresh eyes, I keep trying to relate this to canceled job states, and it seems strange.  Also, the different spellings of canceled/cancelled is annoying.

In short: in favor of dequeue we have: mirrors enqueue, easy to spell, clearly applies to item and not job.

Patch-set: 5
Label: Verified=0