Cache config schema in configloader and drivers
Performance profiling showed that schema creation can be a performance
problem since the greater part of the schema is recreated on every
instantiation of the ParseContext.

This is amplified by having a large number of config files in the zuul.d
directory since for every file there will be call to
TenantParser.parseConfig() which in turn creates the  ParseContext.

Change-Id: I3785cb971426d3c6a9c2aa414dcf042ac36d1b81