Update patch set 1
Patch Set 1: Code-Review-1

I think we want to keep the tenant in the api path always (the one exception is the global info path, but there are specific reasons for that).  It's designed to be whitelabeled, but standard redirection rules won't work if we start moving things up.  This may also impact authorization rules.

Patch-set: 1
Label: Code-Review=-1