Update patch set 1
Patch Set 1:

> I think we want to keep the tenant in the api path always (the one
 > exception is the global info path, but there are specific reasons
 > for that).  It's designed to be whitelabeled, but standard
 > redirection rules won't work if we start moving things up.

There's already /api/info, /api/connections, /api/tenants and /api/user/authorizations (granted, the last one might need a rename anyway), and removing the tenant from endpoints for autohold actions by id would align the REST API with the RPC. Ultimately this is not a big deal and can be abandoned, but it'd put my OCD at rest! :)

 > This may also impact authorization rules.

I don't think so because we get the tenant from fetching the request by id, then authz rules are applied. To make sure of this, I'm adding testing for this specific case in the next PS.

Patch-set: 1
Label: Verified=0