WIP: Link to previous buildset results when reporting a check to Github
The Github checks API only shows the last/current check run for each
commit, but no history of previous runs. To make those results simply
accesible by the user, this patch adds a link to the description of a
check run that points to the buildset result page in Zuul. To only show
the relevant results, the link automatically filters for change and
pipeline.

While this is only used for the checks API by now, it would also
possible to use it by other reporters as well.

Change-Id: Icb81c68f3012781fee343204303cc8f1b69f9f3a