Update patch set 1
Patch Set 1:

There's a "build history" link on the summary page for every build.  This change would save an extra click, however, if you're looking at build history, it seems likely you'd be starting from the most recent build anyway for comparison.  In order to keep what we present to users in GitHub simple, should we just stick with that and not make this change?

Patch-set: 1
Label: Verified=0