Update patch set 1
Patch Set 1:

IMHO those links have different purposes:

The build history link - on job level - is nice to get the previous build results for a specific job and project combination.

Coming from a Pull Request in Github however, I'm more interested in the previous results that were reported by Zuul for this specific PR. In that case I'm more interested in the list of previous buildset results filtered for this change and I'm not so much interested in single build results. If I want to know more about those I can simply follow the links in the Zuul UI.

Another important point is that we filter the buildset results for the specific change (Github PR). Otherwise, we would get a lot of unrelated buildset results which are not of interest here.

Patch-set: 1
Label: Verified=0