Update patch set 3
Patch Set 3: Code-Review-1

Instead of adding cache decorators everywhere, it seems like really we want some kind of object that represents a ParseContext for a Tenant.  How about we have TenantParser.fromYaml return a new class (TenantParseContext?) that holds both a Tenant and a ParseContext?

Alternatively, we could make TenantParser more stateful, but that might be a larger change.

Patch-set: 3
Label: Code-Review=-1