Update patch set 2
Patch Set 2: Code-Review-1

I'm still concerned about putting too much information in the report back to GitHub.  We've gotten a lot of feedback that that is not desirable.

We currently link from GitHub to the buildset page, right?  How about we put a link on the buildset page to previous bulidsets for the same change?  It would be similar to the "build history" link on the build page, but for buildsets.

Patch-set: 2
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1