Include extra, host, and group variables in job serialization
This serialization is used by the REST API as well as configuration
change detection.  Including these variables allows API users to
see a more complete picture of the job as well as corrects a
malfunction in the job config change detection (a change to a job
with file matcher's host-vars, for example, would not trigger running
the job).

Change-Id: I27b5fd81fc0c90bd13ead6ecdef0a5d3dbcc2e57