Update create-react-scripts to 3.4.2
We need to update so we can remove the nanoid pin, which is
just there because the newer version of nanoid uses a .cjs file.

This patch is because json files don't have comments, so we'll
forget otherwise. We should update as soon as c-r-s 3.4.2 is
released.

Relevant PRs on github:

https://github.com/facebook/create-react-app/pull/8768
https://github.com/octet-stream/form-data/issues/18
https://github.com/ai/nanoid/issues/205

Change-Id: I12c1e137ec089fb5e4defb543869a472da204d36