tutorial: Rework quick-start and prepare for other tutorials
-task files are being added and used by quick-start.
-a setup part is split to another file, remainer of previous designs
 where others tutorials did not execute the quick-start tutorial first.
-quick-start steps are modified and fit more to what a user would do

docker-compose: "latest" is just a tag it does not necessarly point to
the latest version available, for gerrit it's the 3.0.4
and even if it points to the latest avaible stable version it can be
buggy. I had the same issue as [1]
So let's use explicit known version that are working

nodepool: declare 4 static nodes, as the following tutorials will
execute up to 2x2 jobs at the same time

project.config: some steps will require direct push to gerrit

"from_json | to_json | from_json | json_query" in test is due to a bug
in ansible [2]

As other tutorials are going to be added, a new job is introduced to
display all tutorials with the same prefix, zuul-quick-start can be
remove afterwards.

[1] https://github.com/docker-library/mariadb/issues/307
[2] https://github.com/ansible-collections/community.general/issues/320

Change-Id: Id5c669537ff5afc7468352139980ebade167d534