Update patch set 3
Patch Set 3:

Sorry, I don't think my idea was quite right.  How about this revision: we create a TenantParseContext inside ConfigLoader.loadConfig, and also ConfigLoader.createDynamicLayout.  That should make it available in the 2 places we need it (those methods are mutually exclusive, we are only going to call one or the other depending on whether we are making a static or dynamic config).  It would only create a ParseContext once for each tenant in each configuration loading event.

Patch-set: 3