Update patch set 3
Patch Set 3: Code-Review-1

(10 comments)

I like the fixes, and the re-organization to group the tasks according to section in the documetation.  I think we could switch away to using the base ubuntu image too, but we should make as few changes as possible.  Generally speaking, I think we should avoid adding extra complexity to the quick-start in order to support follow-on tutorials.  So let's not add more nodes than we need for this, but instead, teach people how to update the nodepool config later.

Patch-set: 3
Label: Code-Review=-1