web UI: user login with OpenID Connect
If the default realm's authentication driver is set to "OpenIDConnect",
display a "Sign in" button. If the the user is logged in, fetch user
authorizations and add them to the redux store; display the user's name
in the upper right corner along a "Sign out" button. Clicking the
sign out button removes user information from the store (note that it
does not log the user out from the identity provider).

Add some basic documentation explaining how to configure Zuul with Google's
authentication, and with a Keycloak server.

Authentication will be intended to be tenant-scoped; for the time being
only the default realm defined in Zuul's configuration is used for all tenants.

Change-Id: I31e71f2795f3f7c4253d0d5b8ed309bfd7d4f98e