zuul-web: support OPTIONS for protected endpoints
CORS-compliant clients can send a preflight query with the OPTIONS
verb to find out which headers, and verbs are allowed on the endpoint.
Implement a response for endpoints that require the "Authorization"
header.

Also fix a bug in the api/user/authorizations endpoint resulting in
a server error if the client does not send an Authorization header.

Change-Id: I34a206ce1dce870f5ec0b8a594e3ebc792cc5908