GitHub Reporter: Fix `Reviewed-by` in Merge Commit Message
Currently, the user stated in the `Reviewed-by` clause of a merge commit
message is the last user that commented on a pull request. This is only
by chance a user that reviewed a pull request.

This change makes sure we append the correct list of GitHub users to the
`Reviewed-by` clause, i.e., those who are in the list of reviews of the
current change/pull request.

Change-Id: Ic1c8187fce2206607789549f02a1d9576818b23e