Add 'uuid' to 'src_dir' in order to allow parallel jobs for a static node
When using max-parallel-jobs>1 for nodepool static provider, 2+ jobs will
overlap and prepare-workspace will remove files created by an already running
jobs

This patch adds 'uuid' to 'src_dir' which allow jobs to work on their own files

'src_dir' is removed from the model, it's not an intrinsic property of a Ref.
It only matters during execution. 'short_name' is added to project dictionary
and Ref now points to a project's data

Depends-On: Ied32ad2b56ba782148a1733091cde351044ef07a
Depends-On: I29f201332286eb975e522802c6d96cc892ad2cd2
Change-Id: Ib78ce4272995680f37ddede73fe6622bcb5c9443