Update patch set 5
Patch Set 5:

Yes that could be an issue at scale for the CE version. As Pierre-Louis said it seems an option exist to set a common webhook token for a project group: from the GitLab doc:
"""
Webhooks can be used to update an external issue tracker, trigger CI jobs, update a backup mirror, or even deploy to your production server. They are available per project for GitLab Community Edition, and per project and per group for GitLab Enterprise Edition.
"""

The is still the solution to manage the webhook token by the GitLab API to handle and sync them for a given set of project.

There is something called integration on GitLab, but it seems to be not as flexible as a Github app and it requires like a partnership with Gitlab. https://docs.gitlab.com/ee/user/project/integrations/overview.html. Then unsure this could simplify the integration between Zuul and GitLab.

Patch-set: 5
Reviewer: Gerrit User 6889 <6889@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0