REST API, Web UI: implement nodes filtering
When querying the nodes API endpoint, filtering options can be added as
URL parameters. Supported parameters are type, provider, state,
in_state_less_than, in_state_more_than (the latter two let a user filter
nodes by how long in seconds they have been in their current state).

Add filtering to the nodes page in the UI.

Add possibility in TableFilters components to transform filter values
or query parameters back and forth (for example convert a duration
in a human-readable format in a filter, to a duration in seconds in
the REST query).

Change-Id: I7ac0bc5c1daf9626b462c35c7699a16f257a5376