Introduce Patternfly 4
Since Patternfly 4 (PF4) is a complete rewrite of the framework that
doesn't require bootstrap and comes with a new npm package, it's
possible to include both side-by-side in a project.

This change includes the necessary PF4 packages and updates the header,
navbar and global page layout with PF4 components. Once this is done, we
should be able to update the other components step by step to PF4.

Points to keep in mind for the migration phase:
  1. Some Patternfly 3 CSS rules are overridden by Patternfly 4
     wildcards, mostly (re)setting the padding and margin of various
     elements to 0.

     To fix this unwanted behaviour, there is a pf4-migration.css file
     included were we can keep track on those rules and ensure that the
     old padding and margin values are re-applied after Patternfly 4 is
     imported.

Open points:
  - The PF4 NavigationDrawer (used to show the config errors) doesn't
    look like a drawer and fills up the whole page instead. I will fix
    this in a follow-up change to not blow this change up any further.
  - I've switched to the logo.svg because the logo.png file was very
    small. Although I really like to larger size of the logo it feels a
    little lost without the 'ZUUL' next to it. Could somebody change the
    logo.svg file to include the text?

Change-Id: I77b81fa0f97fe718207ba5a506cee300371c693b