WIP: Introduce Patternfly 4
Since Patternfly 4 is a complete rewrite of the framework that doesn't
require bootstrap and comes with a new npm package, it's possible to
include both side-by-side in a project.

However, this has the drawback that some of the Patternfly 3 CSS rules
are overwritten by Patternfly 4 wildcards, mostly (re)setting the
padding and margin of various elements to 0.

To fix this unwanted behaviour, there is a pf4-migration.css file
included were we can ensure that the old padding and margin values are
re-applied after Patternfly 4 is imported.

Change-Id: I77b81fa0f97fe718207ba5a506cee300371c693b