web UI: allow a privileged user to re-enqueue a change
Add a "re-enqueue" button on a Buildset page, displayed only
if the currently logged in user is an admin on the tenant. By
clicking this button, the user can re-enqueue the change with
the same parameters as the buildset.

Also fix incorrect changeId being displayed if the buildset
was built for a rev rather than a change.

Change-Id: I05b6b3deb912b121df8de207944d9ec26e7c92d1