zuul_stream: close streamer when task is skipped
Prior to this change: a skipped ansible task calls the start callback
before executing skipped callbacks. In case of a 'command' or 'shell',
it means that a streamer thread is created and is not terminated immediatly
as skipped callbacks were not calling thread.join().

The following sequence illustrates what is happening:

TASK1 shell
-started: streamer opened, thread T1
-skipped: no action on streamer

TASK2 shell
- started: streamer opened, thread T2
- command finished instantly: _log_streamline returns True, streamer thread
  is finished
- ok callback: join thread T1 will block for 10seconds, join thread T2 instant

The wait time is wrongly attributed to TASK2, this wait time can be avoided
if the thread is managed when skipped callbacks are executed

This change modifies how streamer threads are managed, using 'ip' which is
directly related to a the thread

Change-Id: I143c342c8cc457bb6e0f90886b3394e6f7814703