Update patch set 22
Patch Set 22: Code-Review-1

(9 comments)

Not every comment inline is worthy of a -1 but it seems we've added duplicate files and jobs as well as unused variables. It would be good to make this as simple as possible to avoid confusion.

Patch-set: 22
Reviewer: Gerrit User 4146 <4146@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1