Update patch set 10
Patch Set 10:

Thanks for the feedback.

@Tristan: I've fixed the paddings on the status page. Currently, the pf4-migration.css file mostly contains rules for the status page, but AFAIK that's the most customized component in zuul-web and thus might require some more effort in updating to PF4 anyways.

What exactly do you mean by "the tenant list buttons looks too small"? The tenant selector in the navigation bar should be the same size as everything else in the navigation bar. But maybe I misunderstood that.

@Matthieu:
For now, I would like to keep this initial change as small as possible to get the PF4 migration started. I don't know the details about the table filters, but we should update them when we also update the corresponding tables (as I'm not sure how PF4 table filters would work with PF3 tables).

Patch-set: 10
Label: Verified=0