PF4: Update buildset result page (new layout and styling)
This updates the buildset result page with Patternfly 4 components and
a new design based on inspirations from other CI systems and Patternfly
4 demos and my own ideas.

I chose the buildset result page because it doesn't contain too much
elements und therefore is a good starting point. If you like the
design, I will continue with that also for other pages like the build
result pages or the various "list" pages like builds and buildsets.

Change-Id: I978ee448ddf6e22e4a6ec8211204f694932eaa4e